Crate argmin_observer_paramwriter

Writes parameter vectors to a file during optimization.

See documentation of ParamWriter and ParamWriterFormat for details.

Usage

Add the following line to your dependencies list:

[dependencies]
argmin-observer-paramwriter = "0.1.0"

License

Licensed under either of

Apache License, Version 2.0, (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT License (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Structs

ParamWriter

Write parameter vectors to a file during optimization.

Enums

ParamWriterFormat

Available serializers for ParamWriter.